s/DB/Mongo/
Mostly just change naming, but also lay groundwork for implementation
strategy:

Since all documents have UUID _ids that use a timestamp for their high
end bits, we can naively assume a reasonably uniform distribution
between the lowest and highest _ids in a collection. Based on this
assumption, we'll default to splitting based on range queries on _id.

We'll still present the pkColumn option to clients to allow better
handling of collections that are inserted more chunkily.