Port DBMigrate over to MongoMigrate
Swap out the db bits for mongo bits. Currently only supports
creating input splits on ids, which leads to pretty chunky input
splits. A collection that gets updated on fairly regularly across
time should be less chunky thanks to the timestamp in the MSB of
ObjectIds, but it's going to be far from perfect. A much better
solution would be to try to infer the sharding strategy for a
a particular collection and split in alignment with that strategy.