Fixed calculation of chunk number and size.
Chunk number and chunk size was calculated incorrectly, so that some
database entries were ignored. Notice that might have a huge impact on
the correctness of Your cascading job.

To test this, try a collection with less entries than the configured
chunk size (default is 10). All entries will be ignored.

cascading-dbmigrate's calculations are correct:

https://github.com/Cascading/cascading-dbmigrate/blob/master/src/jvm/cascading/dbmigrate/hadoop/DBInputFormat.java